requirements.txt
beautifulsoup4==4.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-language==1.3.0
googleapis-common-protos==1.6.0
goose3==3.1.6
grpcio==1.23.0
idna==2.8
jieba==0.39
lxml==4.4.1
nltk==3.4.5
Pillow==6.1.0
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3
